analytics/gtm-stream-functions/requirements.txt
google-cloud-pubsub>=2.11.0,<3.0.0
functions-framework==3.*
futures>=3.0.5,<4.0.0